requirements.txt
Django==5.1.7
django-admin-sortable2==2.2.4
django-tinymce==4.1.0
pillow==11.1.0
python-environ==0.4.54
requests==2.32.3
